[Updated] Se cambia de posición el inicio de Analytic y se agrega el changelog.
